Side_Project_TG_bot_Memory/requirements.txt
Telethon~=1.27.0
requests~=2.28.2
openai~=0.27.8
pandas~=1.5.3
tenacity~=8.2.2
anthropic~=0.34.1
numpy~=1.24.2
diskcache~=5.6.3
asyncpg~=0.29.0
psycopg2~=2.9.9
Flask~=2.2.3
bs4~=0.0.1
beautifulsoup4~=4.12.2
retrying-async
python-dotenv
